Move user registration form from Home to DashboardAdmin and refactor modal handling
